requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dash==1.1.1
dash-core-components==1.1.1
dash-dangerously-set-inner-html==0.0.2
dash-html-components==1.0.0
dash-renderer==1.0.0
dash-table==4.1.0
Flask==1.1.1
Flask-Compress==1.4.0
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.4
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysqlclient==1.4.4
numpy==1.17.0
pandas==0.25.0
plotly==4.1.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
retrying==1.3.3
six==1.12.0
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.5
yagmail==0.11.220
zipp==0.5.2
